ruby/rails: Token generator supports non-admin users, readme rework (#126)
* ruby/rails: Token generator supports non-admin users

Along with some other minor improvements: satisfy rubocop, extract
region from host name

* ruby/rails: Rework readmes

A mix of reorganization and new content. The intent is to bring these
documents in line with what we want on the aws-samples repo, rather
than what is intended to be on the docs site. Changes are:
- Contributing.md: Call out that this is internal only documentation.
  Setup instructions are polished.
- Readme.md: Describes how the example uses the plugin to generate
  tokens and modifies the ActiveRecord dialect. Customers should
  refer to this for modifying their own applications
- petclinic/Readme.md: Particular (public) instructions for running
  this example.

* ruby/rails: Use env vars for cluster and user, create .internal dir

The .internal dir will hold things we don't want sync'd do the public
repo.

* ruby/rails: Fix non-admin user name for internal use

* ruby/rails: Scrub 'cloud desktop' reference

And a couple internal setup steps