Remove untested Java samples (#196)
This PR removes the `using_dsql_connector` samples which were added in
#190.

These samples do not have associated workflows, meaning there is nothing
guaranteeing they continue to work for users. Additionally, they overlap
with samples already defined in the project which are better tested.

The DSQL connector functionality will be added to the existing samples
in a future change. Release
[1.1.2](https://github.com/awslabs/aurora-dsql-jdbc-connector/releases/tag/1.1.2)
of the JDBC connector improved the user documentation a fair bit, so the
Javadocs should be sufficient to show advanced use cases which were
previously captured by these code samples. Our aim at the moment is to
keep the samples simple and focused on basic connection establishment
and operations. The advanced use cases in the removed samples seemed to
be making up for missing user documentation which has now been addressed
in https://github.com/awslabs/aurora-dsql-jdbc-connector/pull/26.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT-0 license.