Migrate Python Connector Examples
Adding the examples using the Aurora DSQL Python Connector (with the
smoke tests), since using the connector is the preferred method and
offers a more seamless user experience. This is the promoted default
path on the documentation, and the samples should be consistent with
this. Moved the boto3 examples to an `alternatives`, for use as
needed.

This also offers agentic benefits, as most agents are now able to
leverage a single search space and cross-example contex can stay
intact. It creates stronger retrieval since all the variants sit in the
same space, and agents will quickly be able to locate the best example
since by default, the documentation normally points to the samples
repository.