sequelize: Remove duplicate example code from README.md (#217)
This PR cleans up the Sequelize `README.md` file in preparation for
upcoming changes.

The `README.md` previously contained duplicate source code which needed
to be kept in sync with the example source code. I have stripped this
out to simplify the `README.md`, and allow for easier
edits/restructuring.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT-0 license.