sequelize: Remove default database config (#222)
This PR removes the `database: 'postgres'` config, since it is provided
by the node-postgres connector as a default for DSQL.

This default previously wasn't applied, since Sequelize uses an
`undefined` value for the database config when it is not provided by the
user. The node-postgres connector didn't overwrite this with the default
value until
https://github.com/awslabs/aurora-dsql-nodejs-connector/pull/145. This
change was released in
[0.1.6](https://www.npmjs.com/package/@aws/aurora-dsql-node-postgres-connector/v/0.1.6),
so we can now remove the `database` configuration from the sample.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT-0 license.