sequelize: Remove unused skipIndexes dialect option (#218)
While making changes to the Sequelize code sample, I noticed the
`skipIndexes` option is not defined anywhere in Sequelize that I can
find. I'm not sure how it ended up in the sample but I don't think it
has any impact on the behavior of the code, and should be removed.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT-0 license.