sequelize: Document postgres dialect limitations (#220)
This PR improves the Sequelize `README.md` file based on the results of
experimenting with the `postgres` dialect.

The intent is to provide some more targeted guidance for specific issues
users might encounter. The information from the original doc is
retained, but reorganized in a few places to provide extra explanations
and code snippets.

~~I also removed the `database` property from the example since it was
specifying the default, which we should leave to our connector to
configure for simplicity. I noticed this when reworking the connection
section in the `README.md`.~~ Nevermind this doesn't work, I guess I had
a local `PGDATABASE` set while testing.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT-0 license.